requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
google-genai
Pillow
python-dotenv
fitz
requests
pymupdf
frontend
